socketcan-ci.yml was the only conflict. Keep master's real vcan, ARM32,
and injected-failure jobs and port their dotnet test steps to
Microsoft.Testing.Platform: --solution / --project, --report-trx with
--report-trx-filename, and --filter-trait for Category=FakeOnly.

+++ src/adapters/CanKit.Adapter.SocketCAN/Utils/SocketCanUtils.cs
@@ -94,8 +94,8 @@ internal static class SocketCanUtils
     {
         return new Libc.timeval
         {
-            tv_sec = (int)Math.Floor(t.TotalSeconds),
-            tv_usec = (int)((t - TimeSpan.FromSeconds(Math.Floor(t.TotalSeconds))).TotalMilliseconds * 1000.0)
+            tv_sec = (nint)Math.Floor(t.TotalSeconds),
+            tv_usec = (nint)((t - TimeSpan.FromSeconds(Math.Floor(t.TotalSeconds))).TotalMilliseconds * 1000.0)
         };
     }
